pandas/from-mysql-to-bq/requirements.txt
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
fastparquet==0.4.1
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-bigquery==2.3.1
google-cloud-core==1.4.3
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
idna==2.10
llvmlite==0.34.0
numba==0.51.2
numpy==1.19.4
packaging==20.4
pandas==1.1.4
proto-plus==1.11.0
protobuf==3.13.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==0.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.20
thrift==0.13.0
urllib3==1.25.11
